src/api/requirements.txt
Django==2.2.5
django-model-utils==3.2.0
djangorestframework==3.10.2
djangorestframework-stubs==1.0.0
psycopg2-binary==2.8.3
pytz==2019.2
sqlparse==0.3.0
